+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 1.0.0 (2020-12-13)
+
+
+### Bug Fixes
+
+* **changelog:** 优化changelog ([d531833](https://github.com/Jennifer-Liuu/algorithms/commit/d5318332d237b38a84a333d4cae78acff08d2234))
+
+
+### Features
+
+* **array:** three sum ([ba8ccd4](https://github.com/Jennifer-Liuu/algorithms/commit/ba8ccd415e6e80d1adac5e1e83e7af8cdedf6faa))
+* **array & recall:** 字谜分组&回溯算法 ([5c40e80](https://github.com/Jennifer-Liuu/algorithms/commit/5c40e809d12901aa357f32273bb61b65d1a5d6d5))
+* **changelog:** format changelog ([3fc8dab](https://github.com/Jennifer-Liuu/algorithms/commit/3fc8dab84b4a15a02d95508ed3701d2e9436c0d5))
+* **linklist:** addTwoNumbers ([5121010](https://github.com/Jennifer-Liuu/algorithms/commit/5121010b1c01596e551006a84926a0b7a8ca8ee7))
+* **linklist:** 相交链表 ([53f26a7](https://github.com/Jennifer-Liuu/algorithms/commit/53f26a7dec25cf811e9b97e7faa83c30abb11ab6))
+* **recall:** 子集 ([efa7513](https://github.com/Jennifer-Liuu/algorithms/commit/efa75135531558260a05aff6ae528b6d98c59f4b))
+* **sort:** 颜色分类 ([9e221cb](https://github.com/Jennifer-Liuu/algorithms/commit/9e221cb4fbe1033c73ab2a761fc7e65de729980b))
+* **tree:** 二叉树中序遍历 ([09ee3ed](https://github.com/Jennifer-Liuu/algorithms/commit/09ee3edcc327e32ae10ea2571414c52a6e81582e))
+* **tree:** 填充每个节点的下一个右侧节点指针 ([fdbc546](https://github.com/Jennifer-Liuu/algorithms/commit/fdbc5469f95f2039fa5dec3022120e42a88d07ae))
+
+
+
 # 1.0.0 (2020-03-28)